Refactor : require cycle 문제 해결
