+++ app/feature/auth/components/index.ts
@@ -2,4 +2,3 @@ export * from './AuthContext'
 export * from './EmailLoginForm'
 export * from './EmailSignUpForm'
 export * from './NameInputBottomSheet'
-export * from './UserProfileBottomSheet'
