+++ app/feature/magazines/hooks/useBookmarkStatus.ts
@@ -1,7 +1,7 @@
 import { useState, useEffect, useCallback } from 'react'
 
 import { useAuth } from '@/feature/auth/components'
-import { supabase } from '@/feature/shared'
+import { supabase } from '@/feature/shared/lib'
 
 export function useBookmarkStatus(magazineId: string) {
   const { user } = useAuth()
